requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
requests>=2.31.0
psycopg2-binary>=2.9.9
